Extract components from App.js
- Create InkCanvas and NftyHeader to tidy up App.js
- Improve layout for desktop + mobile
